+++ src/main/java/com/pet/sitter/chat/controller/ChatRoomController.java
@@ -1,10 +1,8 @@
-/*
 package com.pet.sitter.chat.controller;
 
-import com.pet.sitter.chat.dto.ChatRoom;
-import com.pet.sitter.chat.service.ChatService;
+import com.pet.sitter.chat.dto.ChatRoomDTO;
+import com.pet.sitter.chat.repository.ChatRoomRepository;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -16,41 +14,40 @@ import org.springframework.web.bind.annotation.ResponseBody;
 
 import java.util.List;
 
-@Controller
-@Slf4j
 @RequiredArgsConstructor
+@Controller
 @RequestMapping("/chat")
 public class ChatRoomController {
 
-    private final ChatService chatService;
+    private final ChatRoomRepository chatRoomRepository;
 
+    // 채팅 리스트 화면
     @GetMapping("/room")
-    public String room(Model model) {
+    public String rooms(Model model) {
         return "/chat/room";
     }
-
+    // 모든 채팅방 목록 반환
     @GetMapping("/rooms")
     @ResponseBody
-    public List<ChatRoom> rooms() {
-        return chatService.findAllRoom();
+    public List<ChatRoomDTO> room() {
+        return chatRoomRepository.findAllRoom();
     }
-
+    // 채팅방 생성
     @PostMapping("/room")
     @ResponseBody
-    public ChatRoom createRoom(@RequestParam String name) {
-        return chatService.createRoom(name);
+    public ChatRoomDTO createRoom(@RequestParam String name) {
+        return chatRoomRepository.createChatRoom(name);
     }
-
+    // 채팅방 입장 화면
     @GetMapping("/room/enter/{roomId}")
     public String roomDetail(Model model, @PathVariable String roomId) {
         model.addAttribute("roomId", roomId);
-
         return "/chat/roomdetail";
     }
-
+    // 특정 채팅방 조회
     @GetMapping("/room/{roomId}")
     @ResponseBody
-    public ChatRoom roomInfo(@PathVariable String roomId) {
-        return chatService.findRoomById(roomId);
+    public ChatRoomDTO roomInfo(@PathVariable String roomId) {
+        return chatRoomRepository.findRoomById(roomId);
     }
-}*/
+}
