--- src/main/java/com/pet/sitter/chat/dto/ChatMessage.java
@@ -1,25 +1,0 @@
-/*
-package com.pet.sitter.chat.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class ChatMessage {
-
-    public enum MessageType {
-        ENTER, TALK, LEAVE
-    }
-
-    private MessageType type;
-    private String roomId;
-    private String sender;
-    private String message;
-    private String time;
-}
-*/
