+++ src/main/java/com/pet/sitter/chat/dto/ChatMessageDTO.java
@@ -1,0 +1,16 @@
+package com.pet.sitter.chat.dto;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+public class ChatMessageDTO {
+
+    private MessageType type; // 메시지 타입
+    private String roomId; // 방번호
+    private String sender; // 메시지 보낸사람
+    private String message; // 메시지
+}
