--- src/main/java/com/pet/sitter/chat/dto/ChatRoom.java
@@ -1,26 +1,0 @@
-/*
-package com.pet.sitter.chat.dto;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import java.util.UUID;
-
-@Getter
-@Setter
-@NoArgsConstructor
-public class ChatRoom {
-
-    private String roomId;
-    private String roomName;
-
-    public static ChatRoom create(String name) {
-        ChatRoom room = new ChatRoom();
-        room.roomId = UUID.randomUUID().toString();
-        room.roomName = name;
-
-        return room;
-    }
-}
-*/
