+++ src/main/java/com/pet/sitter/chat/dto/MessageType.java
@@ -1,0 +1,5 @@
+package com.pet.sitter.chat.dto;
+
+public enum MessageType {
+    ENTER, TALK, LEAVE;
+}
